feat: CAUANDLIB-63 Implemented new pathing function
Reference code has been updated to show example on how to use it

New method is currently locked behind builder flag and will be set to default in a upcoming major release. Old method will be deprecated in an upcoming major release and phased out afterwards

There will be duplicate code for now as updater mechanism is the same. But this is temporary until the old code is phased out